Clean up Collectible3D
