+++ packages/web/package.json
@@ -62,6 +62,7 @@
     "@emotion/css": "^11.11.2",
     "@emotion/styled": "^11.11.0",
     "@fingerprintjs/fingerprintjs-pro": "3.5.6",
+    "@google/model-viewer": "^3.3.0",
     "@hcaptcha/react-hcaptcha": "0.3.6",
     "@juggle/resize-observer": "^3.3.1",
     "@jup-ag/api": "6.0.6",
